pyproject.toml
[project]
name = "macro-pulse-chile"
version = "0.1.0"
description = "Observatorio macroeconómico local-first de Chile y referentes internacionales."
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "streamlit>=1.32.0",
    "pandas>=2.2.0",
    "plotly>=5.19.0",
    "duckdb>=0.10.0",
    "pydantic>=2.6.0",
    "requests>=2.31.0",
    "python-dotenv>=1.0.1",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "black>=24.2.0",
    "ruff>=0.3.0",
]